requirements.txt
praw
prawcore
python-dotenv
pyyaml
mysql-connector